Remove legacy coop step command surface
Committed-By-Agent: codex
