+++ pkg/cmd/coop.go
@@ -34,10 +34,8 @@ The developer confirms each step before the agent moves on.`,
 	cc.cmd.AddCommand(newCoopStartCmd().cmd)
 	cc.cmd.AddCommand(newCoopJoinCmd().cmd)
 	cc.cmd.AddCommand(newCoopAgentCmd().cmd)
-	cc.cmd.AddCommand(newCoopStepCmd().cmd)
 	cc.cmd.AddCommand(newCoopStatusCmd().cmd)
 	cc.cmd.AddCommand(newCoopStopCmd().cmd)
-	cc.cmd.AddCommand(newCoopNextStepsCmd().cmd)
 	cc.cmd.AddCommand(newCoopRecommendCmd().cmd)
 	cc.cmd.AddCommand(newCoopRecoverCmd().cmd)
 	cc.cmd.AddCommand(newCoopDebugAgentCmd().cmd)
