+++ pkg/coop/blueprints/deploy-stripe-projects.json
@@ -2,7 +2,7 @@
   "id": "deploy-stripe-projects",
   "title": "Deploy with Stripe Projects",
   "description": "Configure and deploy your local project using the Stripe Projects CLI plugin.",
-  "prompt": "The developer wants to configure their local project for deployment using the Stripe Projects CLI plugin. You will walk them through the setup using `stripe projects` commands. Use `stripe projects --help` and `stripe projects <subcommand> --help` to discover available options. All stripe CLI commands output JSON when passed --json. Report command output to the developer via the coop step commands.\n\nIMPORTANT: If any step fails because the account is a claimable sandbox (errors mentioning 'claim', 'unclaimed', or permission issues), tell the developer they need to claim their sandbox first. Show them the claim URL from `stripe coop status` (the claim_url field) or from `stripe whoami`. Once they've claimed it, tell them to re-run the session with: stripe coop start deploy",
+  "prompt": "The developer wants to configure their local project for deployment using the Stripe Projects CLI plugin. You will walk them through the setup using `stripe projects` commands. Use `stripe projects --help` and `stripe projects <subcommand> --help` to discover available options. All stripe CLI commands output JSON when passed --json. Report command output to the developer via the coop agent commands.\n\nIMPORTANT: If any step fails because the account is a claimable sandbox (errors mentioning 'claim', 'unclaimed', or permission issues), tell the developer they need to claim their sandbox first. Show them the claim URL from `stripe coop status` (the claim_url field) or from `stripe whoami`. Once they've claimed it, tell them to re-run the session with: stripe coop start deploy",
   "type": "learning",
   "products": [
     "Projects"
